gma config: Tag constants depending on generation or CPU
Tag all derived config booleans that depend on the generation <genbool>
and those that may depend on the CPU with the generations where that is
the case. For instance `CPU_Ivybridge` can be decided purely based on
the generation unless the generation is Ironlake, thus, it is tagged
<ilkbool>. For non-boolean constants, per generation tags <...var> are
introduced (e.g. <ilkvar>) with similar meaning. This will allow us to
make them variables later, based on the generation.

To ease later parsing, also move all multiline expressions after a
line break.

Change-Id: Iaa50987c51dc62ab2eb8b31e8f223b8a94e8ac12
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/27067
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>