gma i2c: Rework GMBUS reset procedure
Once we tried to use the GMBUS controller with an unconnected pair of
I2C pins, it got stuck and a reset via the SOFTWARE_CLEAR_INTERRUPT
bit didn't suffice to recover. Further tests have shown that we are
able to recover, if we switch to a valid pin pair first and issue an
I2C stop cycle before the reset.

Change-Id: If737ffb35afa309de7746f0c16025b9598f69460
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/32730
Reviewed-by: Angel Pons <th3fanbus@gmail.com>