dp training: Always end with normal output
The DP spec mandates that we return to normal output when we abandon
a failed training, before we disable the output. However, the sequence
suggested by Intel always disables the output, even if we try another
time or another frequency. So we also always return to normal frame
delivery.

On Ivy Bridge, failure to return to normal output resulted in inability
to disable the display pipe and to restart the training. Thus, after a
failed training, further attempts were futile.

Change-Id: Ie1f19f2e1d1af62b402ad1dc41373d3d4437e976
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/32731
Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Tested-by: Nico Huber <nico.h@gmx.de>