gma bdw+ transcoder: Use always-on path for primary pipe on eDP
This was a mistake in abb16d9 (gma hsw transcoder: Choose PDW path for
scaling on DDI A). We only have to do it on HSW, future CPUs support
scaling on the always-on path and don't support the alternative path
anymore.

Tested on Haswell (gfx_test), Broadwell (gfx_test) and Coffee Lake
(gfx_test & coreboot): Scaled picture shows up (instead of a black
screen on BDW+).

Change-Id: I257108fce40b73024f231ca366c0d53b700dc588
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/34628
Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: Nico Huber <nico.h@gmx.de>