gma i2c: Increase timeout, again
Can a timeout be long enough? We still face the same problem as in
68f439d4f1 (gma i2c: Increase timeout to >100ms). User reports have
shown, that the 110ms still were not enough, 200ms would be but to
be safe this time, increase the timeout to 500ms. That may result
in very long delays, but only if the hardware is even more broken
that what we encountered so far.

Change-Id: Ia7b28f5a4ed1cd382a7ebfba5ce1ffe1b7a333a8
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/35450
Tested-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>