gma config: Allow to cache CDClk in variable config state
Add `CDClk` and `Max_CDClk` fields to the variable config state.
They will be used to cache the current hardware state, so that
we don't have to poke registers in every call to Update_Outputs().

Also introduce `CDClk_Range`, which is chosen such that common
dot-clock limits (90% of CDClk) are in range of `Frequency_Type`.

Change-Id: I3d9c956eabcedb2f8299a1642ff0172cd7f54e45
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/35714
Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>