Increase range of our main Frequency_Type
Increase `Frequency_Type` to 1MHz .. 2.5Ghz. We used to adapt it
whenever necessary, e.g. when a new generation supported higher
frequencies. But that got a little annoying. So let's choose a
broader range that will hopefully lower the noise for some time.

1MHz is low enough for VGA resolution at 24p, and 2.5GHz is high
enough for 8K at 60p. That should give us some margin.

Some corner cases required refactoring of some calculations and
proofs. And, admittedly, the calculation in `HW.GFX.GMA.PCH.VGA`
can in theory overflow. We prove absence of runtime errors by
deferring the potential overflow to the modular `Word32` type.
This is not perfect, but overflow or not, we have to expect to
be called with a sane frequency anyway.

Change-Id: I3a159bc11ccadcaba8ad04e126e6feadfd46008e
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/35750
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>