gma g45 config: Limit HDMI rate to 165MHz
Tests with a 3440x1440 pixel display have shown that the dot clock
was chosen much too high (we got either 720x1440 at half the rate
or a completely garbled picture). Turns out, Linux' i915 limits the
clock to the original DVI's 165MHz for 4th gen and earlier, so we'll
do that too. This resulted in a stable picture with 3440x1440 @ 31Hz
(the display even synchronized better on this than the configuration
chosen by i915).

Change-Id: I30cb67feedd776148457101d1ca3e3b8538e672f
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/35779
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>