gma panel config: Turn `internal display type` into a `panel port`
Instead of querying for the type of the internal display (LVDS vs. DP),
make it a choice of the port that has a panel hooked up. As some plat-
forms actually have support for two panels, we prepare for that by
naming it `Panel_1_Port`.

We still need to call into the panel logic, when the respective port
is enabled. This will be done in the next commit.

Change-Id: Id3c78441857d07fb34274b9abd3ae6ea42909ff6
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/38242
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: Nico Huber <nico.h@gmx.de>