gma: Map dummy PTEs for buggy VT-d
Extend Setup_Default_GTT() to add 128 dummy page table entries
after the framebuffer. Apparently the IOMMU may report spurious
type 6 errors if nothing is mapped in this range. Also check for
correct alignment of the start of the framebuffer and that the
additonal PTEs fit into the GTT.

Without such dummy entries, an OS driver might report failed
reads at random addresses. It may or may not look like this:

    DMAR: [DMA Read] Request device [00:02.0] PASID ffffffff fault addr\
    5669eb8000 [fault reason 06] PTE Read access is not set

Change-Id: I1b97c6b44c1ffb37d119541c1c97ffe21d244da8
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/40499
Reviewed-by: Michael Niewöhner
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>