gma pipe_setup skl+: Revise default buffer allocation
To be able to use the second plane as a cursor, we need to make
some space in the buffer allocation. The old allocation was based
on the wrong assumption that we only have 3x 160 buffer blocks
available. Actually, the minimum is set by Apollo Lake, with 508
usable buffer blocks.

For a 3840x2160 framebuffer, we need at least 123 buffer blocks
for the first plane (without accounting for an undocumented work-
around condition, 243 blocks otherwise). We allocate the minimum
for cursors (cursor and second plane with y-tiling) as follows:

* First plane: 150 blocks
* Second plane: 11 blocks
* Cursor plane:  8 blocks

Actually, we will never enable the second plane for a cursor at
the same time as the cursor plane is used, so there is still some
potential savings.

Change-Id: Ie471a8b5ea7f0e94aae0c69b0bf128cce1930d2c
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.sourcearcade.org/c/libgfxinit/+/543
Tested-by: Ada SPARK <gnatbot@sourcearcade.org>
Reviewed-by: Thomas Heijligen <src@posteo.de>