gma: actually enable/disable backlight with new backlight control
In order to enable or disable the backlight on platforms with the new
backlight control, both enable bits in the PWM register and the
PP_CONTROL register need to bet set. Currently, the code only sets one
or the other.

Also, reverse the order in the disable case so that PWM gets disabled
after the backlight enable bit in PP_CONTROL.

Tested successfully on Clevo L141CU (CML).

Change-Id: Ib3827e744e86a9d29e09da3c5fd44d226f0b41b5
Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/48770
Tested-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>