Revert "gma config: Enable Restrictions (No_Elaboration_Code)"
This reverts commit 8a9062a0842c70920ac1c95ba48b7d4f2e1538f0.

It was assumed that pragma No_Elaboration_Code would force the compiler
to avoid elaboration code. However, it turns out that it merely checks
what the compiler did.

Newer compiler versions sometimes choose to add elaboration code where
others didn't. So we have to disable the pragma until we find a way to
force compile-time elaboration.

Change-Id: I1b0328c53add53ea713d1ceec3f24b7751363483
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/55423
Reviewed-by: Angel Pons <th3fanbus@gmail.com>