hw-gfx-dp_info.adb: Fix reading `Aux_RD_Interval`
Starting with DP v1.4, the definition of DPCD address 0xe was updated
to include a new field. Bit 7 now indicates whether a DP receiver has
an Extended Receiver Capability field, so `Aux_RD_Interval` is now in
bits 6:0. This change is backwards compatible with older hardware, as
the defined values for `Aux_RD_Interval` range from 0 to 4, and bit 7
is always zero.

However, libgfxinit does not make sure the value is valid before using
it in calculations, so the `EQ_Delay` function may return an extremely
large value if bit 7 is set. To avoid this, add some masks to properly
obtain the value for `Aux_RD_Interval`.

Change-Id: If22d8dbc0517daeaa043ede30bf69e7a65ab154b
Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/67445
Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Tim Wawrzynczak <inforichland@gmail.com>
Tested-by: Nico Huber <nico.h@gmx.de>