gma tgl: Add port detection
Block TCCOLD during initialization so that Modular FIA registers can be
accessed.  Instead of keeping track of the state of TCCOLD all of the
time (and enabling and re-enabling), this patch just blocks TCCOLD
once during initializaton and leaves it that way.

We don't have any knowledge yet if a TC port is used in legacy DP/HDMI
mode or not, so we always handle hotplug events in both hardware blocks.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I674e7c2dcd7738e3d76266ff36675dbd52fa5ae2
Reviewed-on: https://review.sourcearcade.org/c/libgfxinit/+/470
Tested-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Nico Huber <nico.h@gmx.de>
Tested-by: Ada SPARK <gnatbot@sourcearcade.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>