gma: Re-structure TC port types
The `USBC1_DP/_HDMI' names did not draw an accurate picture. If
they are implemented as legacy ports (for HDMI always the case),
they don't show as USBC in schematics, the OS or anywhere else.
We also couldn't tell if a `USBCx_DP' was implemented as legacy
or USB-C.

We split them by actual, physical connector type now: HDMI, DP,
and USBC (for DP-Alt mode).  The `_TCx' suffixes should provide
an idea where they are connected to a socket or SoC.

Tiger Lake and seemingly Alder Lake use the first 4 TC register
sets only. Hence we keep only 4 respective TC ports for now.

Also, as any of the 3 Combo-PHYs (DP1..3) can be used as eDP, we
map that to the port,  that is used with the first set of panel-
control lines. This should hopefully match most board designs.

Change-Id: If9755fe8e1f9252e6e77e22fe910964ad3c3743b
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.sourcearcade.org/c/libgfxinit/+/456
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: Nico Huber <nico.h@gmx.de>