gma power: Allow to explicitly enable DDI/Aux power
We used to enable power for DP-aux channels quite implicitly by faking
configs that use a specific port.  As Tiger Lake requires us to enable
a DDI port's power late during the modesetting sequence,  we introduce
a new API for such cases, and also use it for the DP-aux case.

Tested on HSW, and BXT & CFL where we enable PW2 explicitly now.

Change-Id: I1fd6348ff4855557166495613c6a181f85a818f4
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.sourcearcade.org/c/libgfxinit/+/461
Tested-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>