gma: Get DPCD 1.1+ displays out of D3
DPCD 1.1 added a field to switch the DP sink into D3 and back. Linux
(and potentially other OS') apparently makes use of it when shutting
down, also for a reboot.  This can lead to display failures in core-
boot on a reboot. So we signal any DPCD 1.1+ display unconditionally
to go into D0 before DP training.

Change-Id: Ieec3813ba99ed3a2c6f701cb34a5e2b226fc14c0
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/c/libgfxinit/+/83596
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Tested-by: Nico Huber <nico.h@gmx.de>