gma tgl: Factor TC ownership/connection handling out
The ownership and connection handling differs for Alder Lake.  It uses
kind of the same procedures with different registers and layouts. Move
things into a sub-package `TC.Ownership',  to allow alternative imple-
mentations.

We have to share some register definitions with the sub-package, hence
move related things into the private part of `TC'.

Change-Id: Iaa6b39bdb250c1af88ba83bf78a82cbdd298245e
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.sourcearcade.org/c/libgfxinit/+/505
Reviewed-by: Thomas Heijligen <src@posteo.de>
Tested-by: Nico Huber <nico.h@gmx.de>